+++ include/materials/lambertian.h
@@ -9,7 +9,8 @@
 #include "materials/material.h"
 #include "textures/texture.h"
 
-class lambertian : public material {
+class lambertian : public material
+{
 public:
     lambertian(Texture *a) : albedo(a) {}
     virtual bool scatter(const ray& r_in, const hit_record& rec, vec3& attenuation, ray& scattered) const;