+++ src/materials/metal.cpp
@@ -5,7 +5,7 @@
 #include "materials/metal.h"
 
 
-metal::metal(const vec3& a, float f) : albedo(a) {
+metal::metal(Texture *a, float f) : albedo(a) {
     if (f < 1)
         fuzz = f;
     else
@@ -15,6 +15,6 @@ metal::metal(const vec3& a, float f) : albedo(a) {
 bool metal::scatter(const ray& r_in, const hit_record& rec, vec3& attenuation, ray& scattered) const {
     vec3 reflected = reflect(unit_vector(r_in.direction()), rec.normal);
     scattered = ray(rec.p, reflected + fuzz * random_in_unit_sphere(), r_in.time());
-    attenuation = albedo;
+    attenuation = albedo->value(rec.u, rec.v, rec.p);
     return (dot(scattered.direction(), rec.normal) > 0.0f);
 }