+++ src/scene_geometry/moving_sphere.cpp
@@ -2,6 +2,7 @@
 // Created by Steve Hwan on 10/8/18.
 //
 
+#include "scene_geometry/sphere.h"  // For the UV math
 #include "scene_geometry/moving_sphere.h"
 
 vec3 moving_sphere::center(float time) const {
@@ -20,6 +21,7 @@ bool moving_sphere::hit(const ray& r, float tmin, float tmax, hit_record& rec) c
         if (temp < tmax && temp > tmin) {
             rec.t = temp;
             rec.p = r.point_at_parameter(rec.t);
+            GetSphereUV((rec.p - center(r.time())) / radius, rec.u, rec.v);
             rec.normal = (rec.p - center(r.time())) / radius;
             rec.mat_ptr = mat_ptr;
             return true;
@@ -29,6 +31,7 @@ bool moving_sphere::hit(const ray& r, float tmin, float tmax, hit_record& rec) c
         if (temp < tmax && temp > tmin) {
             rec.t = temp;
             rec.p = r.point_at_parameter(rec.t);
+            GetSphereUV((rec.p - center(r.time())) / radius, rec.u, rec.v);
             rec.normal = (rec.p - center(r.time())) / radius;
             rec.mat_ptr = mat_ptr;
             return true;